Add Web Extension variant
