+++ client/src/common/localStorageService.ts
