+++ client/src/common/logDocument.ts
