+++ client/webpack.config.js
@@ -2,16 +2,33 @@
 
 'use strict';
 
-const withDefaults = require('../shared.webpack.config');
+const { withDefaults, withDefaultsWeb } = require('../shared.webpack.config');
 const path = require('path');
 
-module.exports = withDefaults({
-	context: path.join(__dirname),
+const nodeConfig = withDefaults({
 	entry: {
-		extension: './src/extension.ts',
+		extension: './client/src/node/extension.ts',
 	},
 	output: {
-		filename: 'extension.js',
-		path: path.join(__dirname, 'out'),
+		filename: '[name].js',
+		path: path.join(__dirname, 'dist/node'),
+		libraryTarget: "commonjs"
 	},
 });
+
+const webConfig = withDefaultsWeb({
+	context: path.join(__dirname),
+	mode: 'none', // this leaves the source code as close as possible to the original (when packaging we set this to 'production')
+	target: 'webworker', // extensions run in a webworker context
+	entry: {
+		extension: './src/web/extension.ts'
+	},
+	output: {
+		filename: '[name].js',
+		path: path.join(__dirname, 'dist/web'),
+		libraryTarget: 'commonjs',
+		devtoolModuleFilenameTemplate: '[absolute-resource-path]'
+	}
+});
+
+module.exports = [nodeConfig, webConfig];
