+++ server/src/common/index.ts
@@ -1,5 +1,5 @@
 import { basename } from 'path';
-import { Location, Range, Diagnostic } from 'vscode-languageserver/node';
+import { Location, Range, Diagnostic } from 'vscode-languageserver';
 
 import { CaseInsensitiveMap, ReadonlyCaseInsensitiveMap, normalizeUri, mapToUnionedCaseInsensitiveMap } from './utilities';
 import { uriIsStartupFile } from './language';
