+++ server/src/common/language.ts
@@ -1,4 +1,4 @@
-import { CompletionItem, CompletionItemKind } from 'vscode-languageserver/node';
+import { CompletionItem, CompletionItemKind } from 'vscode-languageserver/browser';
 
 import { getFilenameFromUri, extractToMatchingDelimiter } from './utilities';
 import { ReadonlyAchievementIndex } from "./index";
