+++ server/src/common/parser.ts
@@ -1,4 +1,4 @@
-import { Range, Location, Position, Diagnostic, DiagnosticSeverity } from 'vscode-languageserver/node';
+import { Range, Location, Position, Diagnostic, DiagnosticSeverity } from 'vscode-languageserver/browser';
 import { TextDocument } from 'vscode-languageserver-textdocument';
 
 import {
