+++ server/tsconfig.json
@@ -6,7 +6,8 @@
 		"sourceMap": true,
 		"strict": true,
 		"outDir": "out",
-		"rootDir": "src"
+		"rootDir": "src",
+		"lib": ["ES6", "webworker"]
 	},
 	"include": ["src"],
 	"exclude": ["node_modules", ".vscode-test"]
